requirements.txt
# TerraTorch Building Segmentation - Dependencies

# Core
terratorch>=0.5.0
torch>=2.0.0
torchvision>=0.15.0
lightning>=2.0.0
torchgeo>=0.5.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0

# Data & Metrics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
jsonargparse[signatures]>=4.20.0

# Utilities
tqdm>=4.65.0
